Fixed the searchbar not displaying the correct results | Fixed the add to order button not adding to order box
