+++ src/components/SearchBar/SearchBar.js
@@ -1,16 +1,6 @@
-import { useState } from 'react';
-import './SearchBar.css';
+import "./SearchBar.css";
 
-const SearchBar = ({ items, setFilteredItems }) => {
-  const [input, setInput] = useState('');
-
-  const handleChange = (e) => {
-    setInput(e.target.value);
-    const filteredItems = items.filter((item) => {
-      return item.name.toLowerCase().includes(input.toLowerCase());
-    });
-    setFilteredItems(filteredItems);
-  };
+const SearchBar = ({ input, handleChange }) => {
   return (
     <div className="SearchBar">
       <input
